# docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntryDTO:
    properties:
      correlation_id:
        type: string
      created_at:
        type: string
      event_id:
        type: string
      event_type:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      payload:
        additionalProperties: {}
        type: object
      source:
        type: string
      tenant_id:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntrySummaryDTO:
    properties:
      correlation_id:
        type: string
      event_type:
        type: string
      id:
        type: string
      source:
        type: string
      tenant_id:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_audit_application_dto.GetActorActivityResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntrySummaryDTO'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total_count:
        type: integer
      user_id:
        type: string
    type: object
  github_com_0xsj_nexus-go_internal_audit_application_dto.GetEntryResponse:
    properties:
      entry:
        $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntryDTO'
    type: object
  github_com_0xsj_nexus-go_internal_audit_application_dto.GetResourceTrailResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntrySummaryDTO'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      resource_id:
        type: string
      resource_type:
        type: string
      total_count:
        type: integer
    type: object
  github_com_0xsj_nexus-go_internal_audit_application_dto.ListEntriesResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.AuditEntrySummaryDTO'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total_count:
        type: integer
    type: object
...
  internal_audit_interface_http_v1.ErrorResponse:
    description: Error response returned by the API
    properties:
      code:
        example: ENTRY_NOT_FOUND
        type: string
      message:
        example: audit entry not found
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Multi-tenant API built with hexagonal architecture. Provides identity
    management, tenant management, email templates, and more.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Hexagonal Go API
  version: "1.0"
paths:
...
  /api/v1/audit/actors/{userID}:
    get:
      description: Retrieves all audit entries performed by a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Filter by tenant ID
        in: query
        name: tenant_id
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by source domain
        in: query
        name: source
        type: string
      - description: Filter from timestamp (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to timestamp (RFC3339)
        in: query
        name: to
        type: string
      - description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Actor activity
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.GetActorActivityResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get actor activity
      tags:
      - audit
  /api/v1/audit/entries:
    get:
      description: Retrieves audit entries with optional filtering and pagination
      parameters:
      - description: Filter by tenant ID
        in: query
        name: tenant_id
        type: string
      - description: Filter by user ID (actor)
        in: query
        name: user_id
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by source domain
        in: query
        name: source
        type: string
      - description: Filter by correlation ID
        in: query
        name: correlation_id
        type: string
      - description: Filter from timestamp (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to timestamp (RFC3339)
        in: query
        name: to
        type: string
      - description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of audit entries
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.ListEntriesResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit entries
      tags:
      - audit
  /api/v1/audit/entries/{id}:
    get:
      description: Retrieves a single audit entry by its ID
      parameters:
      - description: Audit entry ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit entry found
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.GetEntryResponse'
        "400":
          description: Invalid entry ID
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
        "404":
          description: Entry not found
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get audit entry
      tags:
      - audit
...
  /api/v1/audit/resources/{type}/{id}:
    get:
      description: Retrieves all audit entries for a specific resource
      parameters:
      - description: Resource type (e.g., user, tenant, flag)
        in: path
        name: type
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by tenant ID
        in: query
        name: tenant_id
        type: string
      - description: Filter from timestamp (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to timestamp (RFC3339)
        in: query
        name: to
        type: string
      - description: Page size (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Resource audit trail
          schema:
            $ref: '#/definitions/github_com_0xsj_nexus-go_internal_audit_application_dto.GetResourceTrailResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_audit_interface_http_v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get resource audit trail
      tags:
      - audit
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...